Use 32-bit IDs everywhere to accommodate 18-bit shift

+++ inc/Platform/CAN/Frame.hpp
@@ -8,14 +8,14 @@ namespace CAN {
     class Frame {
     public:
         static constexpr uint8_t MaxDataLength = 8;
-        uint16_t id = 0;
+        uint32_t id = 0;
         etl::array<uint8_t, MaxDataLength> data = {};
 
         Frame() = default;
 
-        Frame(uint16_t id) : id(id) {};
+        Frame(uint32_t id) : id(id) {};
 
-        Frame(uint16_t id, const etl::array<uint8_t, MaxDataLength> &data) : id(id), data(data) {};
+        Frame(uint32_t id, const etl::array<uint8_t, MaxDataLength> &data) : id(id), data(data) {};
 
         inline void empty() {
             id = 0;
